Renaming
Config -> ConfigReader
Make const available as pyaerocom.config
