+++ tests/colocation/test_colocated_data.py
@@ -9,7 +9,7 @@ from numpy.typing import ArrayLike
 from pydantic import ValidationError
 
 from pyaerocom import ColocatedData
-from pyaerocom.config import ALL_REGION_NAME
+from pyaerocom.config_reader import ALL_REGION_NAME
 from pyaerocom.exceptions import DataCoverageError, DataDimensionError
 from tests.fixtures.collocated_data import EXAMPLE_FILE
