+++ tests/fixtures/aeroval/cfg_test_exp2.py
@@ -1,4 +1,4 @@
-from pyaerocom.config import ALL_REGION_NAME
+from pyaerocom.config_reader import ALL_REGION_NAME
 
 MODELS = {
     "TM5-AP3-CTRL": dict(
