+++ tests/fixtures/tm5.py
@@ -7,7 +7,7 @@ import pytest
 from xarray import load_dataarray
 
 from pyaerocom import ColocatedData
-from pyaerocom.config import ALL_REGION_NAME
+from pyaerocom.config_reader import ALL_REGION_NAME
 from pyaerocom.griddeddata import GriddedData
 
 from .data_access import DataForTests
