+++ tests/test_filter.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from pyaerocom.config import ALL_REGION_NAME
+from pyaerocom.config_reader import ALL_REGION_NAME
 from pyaerocom.filter import Filter
